% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_correlation_and_regression.R
\name{calculate_correlation_and_regression}
\alias{calculate_correlation_and_regression}
\title{Calculate Correlation Coefficient and Linear Regression Model}
\usage{
calculate_correlation_and_regression(data, x_var, y_var)
}
\arguments{
\item{data}{The input data frame.}

\item{x_col}{The x-axis variable.}

\item{y_col}{The y-axis variable}
}
\value{
The correlation coefficient and linear regression model.
}
\description{
Calculate the correlation coefficient and linear regression between two variables.
}
\examples{
exoplanets <- read.csv("PS_2023.10.04_12.36.20.csv", skip = 290)
calculate_correlation_and_regression(exoplanets, "pl_radj", "st_rad")

}
