% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_barplot.R
\name{create_barplot}
\alias{create_barplot}
\title{Create a Bar Plot}
\usage{
create_barplot(data, x_col, y_col, title, x_label, y_label)
}
\arguments{
\item{data}{The input data frame.}

\item{x_col}{The x-axis data column.}

\item{y_col}{The y-axis data column.}

\item{title}{The title for the plot.}

\item{x_label}{The label for the x-axis.}

\item{y_label}{The label for the y-axis.}
}
\value{
A ggplot2 bar plot.
}
\description{
Create a bar plot using ggplot2 with specified data and labels.
}
\examples{
exoplanets <- read.csv("PS_2023.10.04_12.36.20.csv", skip = 290)
create_barplot(exoplanets, discoverymethod, count, "Exoplanet Discoveries by Discovery Method", "Discovery Method", "Number of Discoveries")

}
