% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lineplot.R
\name{create_line_plot}
\alias{create_line_plot}
\title{Create a Line Plot}
\usage{
create_line_plot(data, x_col, y_col, title, x_label, y_label)
}
\arguments{
\item{data}{The input data frame.}

\item{x_col}{The x-axis data column.}

\item{y_col}{The y-axis data column.}

\item{title}{The title for the plot.}

\item{x_label}{The label for the x-axis.}

\item{y_label}{The label for the y-axis.}
}
\value{
A ggplot2 line plot.
}
\description{
Create a line plot using ggplot2 with specified data and labels.
}
\examples{
exoplanets <- read.csv("PS_2023.10.04_12.36.20.csv", skip = 290)
create_line_plot(exoplanets, disc_year, count, "Number of Exoplanet Discoveries Over Time", "Discovery Year", "Number of Discoveries")

}
