% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scatterplot.R
\name{create_scatterplot}
\alias{create_scatterplot}
\title{Create a Scatter Plot}
\usage{
create_scatterplot(
  data,
  x_col,
  y_col,
  title,
  x_label,
  y_label,
  y_max_limit = NULL
)
}
\arguments{
\item{data}{The input data frame.}

\item{x_col}{The x-axis data column.}

\item{y_col}{The y-axis data column.}

\item{title}{The title for the plot.}

\item{x_label}{The label for the x-axis.}

\item{y_label}{The label for the y-axis.}
}
\value{
A ggplot2 scatter plot.
}
\description{
Create a scatter plot using ggplot2 with specified data and labels.
}
\examples{
exoplanets <- read.csv("PS_2023.10.04_12.36.20.csv", skip = 290)
create_scatterplot(exoplanets, pl_radj, st_rad, "Comparison of Planet Radius and Host Star Radius", "Planet Radius (in Jupiter Radii)", "Host Star Radius (in Solar Radii)", y_max_limit = 6)

}
